Criada tratativas de erros

+++ src/hooks/useCart.tsx
@@ -103,6 +103,8 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
             toast.error('Quantidade solicitada fora de estoque');
           }
         }
+      } else {
+        throw new Error();
       }
       localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
     } catch {
@@ -112,11 +114,14 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
 
   const removeProduct = (productId: number) => {
     try {
-      const updatedCart = cart.filter((item) => item.id !== productId);
-      setCart(updatedCart);
-      localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+      if (existProductInCart(productId)) {
+        const updatedCart = cart.filter((item) => item.id !== productId);
+        setCart(updatedCart);
+        localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+      } else {
+        throw new Error();
+      }
     } catch {
-      // TODO
       toast.error('Erro na remoção do produto');
     }
   };
@@ -128,24 +133,31 @@ export function CartProvider({ children }: CartProviderProps): JSX.Element {
   }: UpdateProductAmount) => {
     try {
       // TODO
-      const updatedCart = cart.map((item) => {
-        if (item.id === productId) {
-          if (operator === 'increment') {
-            if (checkAvaliableStock(productId, item.amount + 1)) {
-              item.amount += 1;
+      if (existProductInCart(productId)) {
+        const updatedCart = cart.map((item) => {
+          if (item.id === productId) {
+            if (operator === 'increment') {
+              if (checkAvaliableStock(productId, item.amount + 1)) {
+                item.amount += 1;
+              } else {
+                toast.error('Quantidade solicitada fora de estoque');
+              }
             } else {
-              toast.error('Quantidade solicitada fora de estoque');
+              if (item.amount - 1 < 1) {
+                throw new Error();
+              } else {
+                item.amount -= 1;
+              }
             }
-          } else {
-            item.amount -= 1;
           }
-        }
-        return item;
-      });
-      setCart(updatedCart);
-      localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+          return item;
+        });
+        setCart(updatedCart);
+        localStorage.setItem('@RocketShoes:cart', JSON.stringify(cart));
+      } else {
+        throw new Error();
+      }
     } catch {
-      // TODO
       toast.error('Erro na alteração de quantidade do produto');
     }
   };
